+++ src/Symfony/Component/Mailer/Bridge/Amazon/Tests/Transport/SesApiAsyncAwsTransportTest.php
@@ -121,7 +121,8 @@ class SesApiAsyncAwsTransportTest extends TestCase
 
         $message = $transport->send($mail);
 
-        $this->assertSame('foobar', $message->getMessageId());
+        $this->assertSame('foobar', $message->getInfo('id'));
+        $this->assertNotSame('foobar', $message->getMessageId());
     }
 
     public function testSendWithNonAsciiCustomHeaders()